agent.py
import logging

from dotenv import load_dotenv
from livekit.agents import (
    AutoSubscribe,
    JobContext,
    JobProcess,
    WorkerOptions,
    cli,
    llm as LLLLM,
    metrics,
)
from livekit.agents.pipeline import VoicePipelineAgent
from livekit.plugins import cartesia, deepgram, silero, turn_detector
from livekit.plugins.openai import stt, llm


load_dotenv(dotenv_path=".env.local")
logger = logging.getLogger("voice-agent")

def prewarm(proc: JobProcess):
    proc.userdata["vad"] = silero.VAD.load()


async def entrypoint(ctx: JobContext):
    initial_ctx = LLLLM.ChatContext().append(
        role="system",
        text=(
            
            """You should use short and concise responses, and avoiding usage of unpronouncable punctuation. 
                You are Urmi, a travel agent. A user will be calling you regarding the trip they want to plan.
                Warm Greeting:
                Start with a warm and friendly greeting to make the user feel at ease.
                Ask About the Duration:
                Inquire about the number of days they're planning for their trip in a casual manner.
                Ask:
                "To kick things off, how many days are you thinking about for your getaway? Just a rough idea is perfect!"
                Inquire About Their Plans:
                Ask more about their specific plans and interests for the trip, encouraging them to share.
                Ask:
                "Awesome! What kind of experiences or activities are you hoping to include? I can't wait to hear all about it!"
                Listen Actively:
                As they share, listen closely and show genuine interest in their ideas.
                Respond with:
                "That sounds really exciting! I love where this is going. Let's dive a little deeper into what you're imagining!"
                Based on user preferences. Recommend them hotels and other things they requireZ by yourself.
                Discuss Key Elements:
                Guide the conversation towards important trip elements like dining, attractions, and activities, keeping the tone light and friendly.
                Restaurants:
                Ask: "When it comes to food, do you prefer cozy little spots, fancy dinners, or maybe trying out some local gems? I've got some tasty recommendations!"
                Places to Visit:
                Ask: "What kind of sights are on your radar? Do you have any must-see places in mind, or do you want me to share some hidden treasures?"
                Leisure Activities:
                Ask: "What do you feel like doing most? Are you in the mood for some chill beach time, or maybe some fun adventures? Let me know what excites you!"
                Offer Alternatives:
                Provide a few options for each suggestion, keeping it conversational and engaging.
                Example:
                "If you're up for some outdoor fun, I can suggest a gorgeous hiking trail, an easygoing bike ride, or maybe even a guided tour. Which one are you leaning towards?"
                Assure Booking Assistance:
                Let the user know you're here to help with their bookings and accommodations, creating a sense of partnership.
                Say:
                "Once we nail down the details, I'll be right here to help you with all the bookings—hotels, restaurants, activities—whatever you need!"
                Lead the Conversation When Necessary:
                Take the lead only when it feels right, ensuring the user feels in control.
                Suggest:
                "If you think it would help, we could sketch out a daily plan together. Just let me know, and I'm happy to guide you through it!"
                Final Thoughts:
                As you wrap up the discussion, summarize everything lightly.
                Say:
                "Wow, we've come up with some really fun ideas! I can help put everything together into a travel plan if you'd like!"
                Wrap Up Positively:
                Conclude on an upbeat note, expressing enthusiasm for their plans.
                Example:
                "I'm so excited for you! This trip is going to be amazing. If anything else comes to mind or you have more questions, just reach out anytime!"

                General Tone and Behavior:
                Keep the tone friendly, warm, and approachable to foster a comfortable and engaging conversation.
                Show genuine curiosity and excitement about the user's plans, encouraging them to share more.
                Use informal language to create a more relaxed and relatable atmosphere.
                Emphasize partnership and collaboration throughout the planning process.

            """
            
        ),
    )

    logger.info(f"connecting to room {ctx.room.name}")
    await ctx.connect(auto_subscribe=AutoSubscribe.AUDIO_ONLY)

    # Wait for the first participant to connect
    participant = await ctx.wait_for_participant()
    logger.info(f"starting voice assistant for participant {participant.identity}")

    # This project is configured to use Deepgram STT, OpenAI LLM and Cartesia TTS plugins
    # Other great providers exist like Cerebras, ElevenLabs, Groq, Play.ht, Rime, and more
    # Learn more and pick the best one for your app:
    # https://docs.livekit.io/agents/plugins
    agent = VoicePipelineAgent(
        vad=ctx.proc.userdata["vad"],
        stt=stt.STT.with_groq(model="whisper-large-v3-turbo", language="en"),
        llm=llm.LLM.with_groq(model="llama3-8b-8192", temperature=0.8,),
        tts=cartesia.TTS(),
        turn_detector=turn_detector.EOUModel(),
        # minimum delay for endpointing, used when turn detector believes the user is done with their turn
        min_endpointing_delay=0.5,
        # maximum delay for endpointing, used when turn detector does not believe the user is done with their turn
        max_endpointing_delay=5.0,
        chat_ctx=initial_ctx,
    )

    usage_collector = metrics.UsageCollector()

    @agent.on("metrics_collected")
    def on_metrics_collected(agent_metrics: metrics.AgentMetrics):
        metrics.log_metrics(agent_metrics)
        usage_collector.collect(agent_metrics)

    agent.start(ctx.room, participant)

    # The agent should be polite and greet the user when it joins :)
    await agent.say("Hey there! I’m Urmi, and I’m super excited to help you plan your trip! What fantastic adventure do you have in mind?", allow_interruptions=True)


if __name__ == "__main__":
    
    cli.run_app(
        WorkerOptions(
            entrypoint_fnc=entrypoint,
            prewarm_fnc=prewarm,
            # giving this agent a name of: "inbound-agent"
            agent_name="inbound-agent",
        ),
    )



